requirements.txt
fastapi
httpx
uvicorn
python-dotenv
cachetools
